# cmd/backend/docs/backend_swagger.yaml
definitions:
...
  trx-project_internal_model.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      status:
        description: '1: 活跃, 0: 禁用'
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  trx-project_pkg_response.Response:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        description: 数据
      message:
        description: 消息
        type: string
      trace_id:
        description: 追踪ID（可选）
        type: string
    type: object
info:
  contact: {}
paths:
...
  /admin/statistics/users:
    get:
      consumes:
      - application/json
      description: 获取用户相关的统计数据，如总数、活跃数等
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取统计信息
          schema:
            allOf:
            - $ref: '#/definitions/trx-project_pkg_response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户统计信息（后台）
      tags:
      - 统计信息
  /admin/users:
    get:
      consumes:
      - application/json
      description: 分页获取用户列表，支持状态筛选和关键词搜索
      parameters:
      - default: 1
        description: 页码，默认1
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量，默认10
        in: query
        name: page_size
        type: integer
      - description: 用户状态筛选
        in: query
        name: status
        type: string
      - description: 关键词搜索（用户名或邮箱）
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取用户列表
          schema:
            allOf:
            - $ref: '#/definitions/trx-project_pkg_response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户列表（后台）
      tags:
      - 用户管理
  /admin/users/{id}:
    delete:
      consumes:
      - application/json
      description: 根据用户ID删除用户（软删除）
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "400":
          description: 无效的用户ID
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 删除用户（后台）
      tags:
      - 用户管理
    get:
      consumes:
      - application/json
      description: 根据用户ID获取用户的详细信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取用户详情
          schema:
            allOf:
            - $ref: '#/definitions/trx-project_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/trx-project_internal_model.User'
              type: object
        "400":
          description: 无效的用户ID
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户详情（后台）
      tags:
      - 用户管理
...
  /admin/users/{id}/reset-password:
    post:
      consumes:
      - application/json
      description: 管理员重置指定用户的密码
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 新密码
        in: body
        name: request
        required: true
        schema:
          properties:
            new_password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 重置成功
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 重置用户密码（后台）
      tags:
      - 用户管理
...
  /admin/users/{id}/status:
    put:
      consumes:
      - application/json
      description: 更新用户的状态（启用/禁用）
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态信息，1-启用 0-禁用
        in: body
        name: request
        required: true
        schema:
          properties:
            status:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/trx-project_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 更新用户状态（后台）
      tags:
      - 用户管理
swagger: "2.0"